error handling nd logic fix

+++ src/controllers/CourseController/createcourse.ts
@@ -2,7 +2,7 @@ import User from "../../models/user.model";
 import Course from "../../models/course.model";
 import Topic from "../../models/topic.model";
 import AIGenerator from "../../utils/services/aigenerator";
-import { AuthError, ClientError } from "../../utils/handlers/error";
+import { AuthError, ClientError, CustomError } from "../../utils/handlers/error";
 import { FileService } from "../../utils/services/parseFile";
 import fs from "fs";
 
@@ -15,18 +15,20 @@ class CourseCreationService {
     this.FileService = new FileService();
   }
   private async createCourseStructure(courseTitle: string, userId: string) {
-    const courseDescription = await this.AIGenerator.generateCourseDescription(
-      courseTitle
-    );
-    const courseCategory = await this.AIGenerator.generateCoursecategory(
-      courseTitle
-    );
-    const courseImage = await this.AIGenerator.generateCourseImage(courseTitle);
+    let courseError: CustomError | null = null;
+    let newCourse = null
     const user = await User.findById(userId);
     if (!user) {
-      throw new AuthError("User does not exist, cannot create course");
+      courseError = new AuthError("User does not exist, cannot create course");
     } else {
-      const newCourse = new Course({
+      const courseDescription = await this.AIGenerator.generateCourseDescription(
+        courseTitle
+      );
+      const courseCategory = await this.AIGenerator.generateCoursecategory(
+        courseTitle
+      );
+      const courseImage = await this.AIGenerator.generateCourseImage(courseTitle);
+      newCourse = new Course({
         title: courseTitle,
         description: courseDescription,
         category: courseCategory,
@@ -34,8 +36,8 @@ class CourseCreationService {
         user: user!._id,
       });
       await newCourse.save();
-      return newCourse;
     }
+    return {newCourse, courseError}
   }
   private async createTopics(courseId: string, courseTitle: string, files?: any, subtopics?: string){
     if(files && subtopics){
@@ -57,7 +59,18 @@ class CourseCreationService {
     } else if (!files && subtopics){
 
     } else {
-
+      const newTopics = await this.AIGenerator.generateTopics(
+        courseTitle
+      );
+      for (const topic of newTopics) {
+        const topicPosition = newTopics.indexOf(topic) + 1;
+        await new Topic({
+          title: topic,
+          position: topicPosition,
+          course: courseId,
+          inProgress: topicPosition === 1 ? true : false,
+        }).save();
+      }
     }
   }
   async createCourse(
@@ -66,44 +79,56 @@ class CourseCreationService {
     files?: any,
     subtopics?: string,
   ) {
+    let error: CustomError | null = null;
+    let courseId: string = "";
     if (title.length < 1) {
-      throw new ClientError(
+      error =  new ClientError(
         "Invalid course title, please enter a valid title."
       );
     } else {
-      const newCourse = await this.createCourseStructure(title, userId);
-      if (files && !subtopics) {
-        // Add files to course and generate course topics
-        const filePaths = files.map((file: any) => file.path);
-        const fileData = await Promise.all(
-          filePaths.map(async (filePath: string) => {
-            const data = await this.FileService.parseFile(filePath);
-            return data;
-          })
-        );
-
-        filePaths.forEach((filePath: string) => {
-          fs.unlinkSync(filePath);
-        });
-
-        // Add files to course
-        await Course.updateOne(
-          { _id: newCourse._id },
-          { $set: { courseFiles: fileData } }
-        );
-        await newCourse.save()
-        await this.createTopics(newCourse._id, newCourse.title, newCourse.courseFiles.join(" "))
-
-      } else if (files && subtopics) {
-        // Add files to course and include the subtpics in implementation
-      } else if (!files && subtopics) {
-        // Implement topics creation from just subtopics provided
-      } else {
-        // Auto-generate topics
-        await this.createTopics(newCourse._id, newCourse.title)
+      const {newCourse, courseError} = await this.createCourseStructure(title, userId);
+      if(courseError){
+        error = courseError
+      } else if(newCourse) {
+        courseId = newCourse._id
+        if (files && !subtopics) {
+          // Add files to course and generate course topics
+          const filePaths = files.map((file: any) => file.path);
+          const fileData = await Promise.all(
+            filePaths.map(async (filePath: string) => {
+              const {data, fileError} = await this.FileService.parseFile(filePath);
+              if(fileError){
+                error = fileError
+              }
+              return data;
+            })
+          );
+  
+          filePaths.forEach((filePath: string) => {
+            fs.unlinkSync(filePath);
+          });
+  
+          // Add files to course
+          await Course.updateOne(
+            { _id: newCourse._id },
+            { $set: { courseFiles: fileData } }
+          );
+          await newCourse.save()
+          await this.createTopics(newCourse._id, newCourse.title, newCourse.courseFiles.join(" "))
+  
+        } else if (files && subtopics) {
+          // Add files to course and include the subtpics in implementation
+        } else if (!files && subtopics) {
+          // Implement topics creation from just subtopics provided
+        } else {
+          // Auto-generate topics
+          await this.createTopics(newCourse._id, newCourse.title)
+        }
       }
-      return newCourse._id;
+      
+      
     }
+    return { courseId, error }
   }
